+++ projects/fhi-angular-highcharts/src/lib/constants-and-enums/diagram-type-ids.ts
@@ -2,11 +2,13 @@ export enum DiagramTypeIds {
   bar,
   barStacked,
   column,
+  columnAndLine,
   columnStacked,
   line,
-  map,
+  map, // TODO: remove in v5
   mapFylker,
   mapFylker2019,
+  mapFylker2023,
   pie,
   table,
 }
@@ -15,11 +17,12 @@ export const DiagramTypeIdValues = {
   bar: DiagramTypeIds[DiagramTypeIds.bar],
   barStacked: DiagramTypeIds[DiagramTypeIds.barStacked],
   column: DiagramTypeIds[DiagramTypeIds.column],
+  columnAndLine: DiagramTypeIds[DiagramTypeIds.columnAndLine],
   columnStacked: DiagramTypeIds[DiagramTypeIds.columnStacked],
   line: DiagramTypeIds[DiagramTypeIds.line],
-  map: DiagramTypeIds[DiagramTypeIds.map],
   mapFylker: DiagramTypeIds[DiagramTypeIds.mapFylker],
   mapFylker2019: DiagramTypeIds[DiagramTypeIds.mapFylker2019],
+  mapFylker2023: DiagramTypeIds[DiagramTypeIds.mapFylker2023],
   pie: DiagramTypeIds[DiagramTypeIds.pie],
   table: DiagramTypeIds[DiagramTypeIds.table],
 };
@@ -28,6 +31,7 @@ export enum ChartTypeIds {
   bar,
   barStacked,
   column,
+  columnAndLine,
   columnStacked,
   line,
   pie,
@@ -37,9 +41,11 @@ export enum ChartTypeIds {
 export enum MapTypeIds {
   mapFylker,
   mapFylker2019,
+  mapFylker2023,
 }
 
 export const MapTypeIdValues = {
   mapFylker: MapTypeIds[MapTypeIds.mapFylker],
   mapFylker2019: MapTypeIds[MapTypeIds.mapFylker2019],
+  mapFylker2023: MapTypeIds[MapTypeIds.mapFylker2023],
 };