+++ projects/fhi-angular-highcharts/src/lib/models/fhi-diagram-unit.model.ts
@@ -14,4 +14,5 @@ export interface FhiDiagramUnit {
   label: string;
   symbol?: string;
   position?: keyof typeof UnitSymbolPosition;
+  yAxis?: number;
 }