+++ projects/fhi-angular-highcharts/src/lib/models/metadata-for-serie.model.ts
@@ -1,0 +1,5 @@
+export interface MetadataForSerie {
+  decimals: number;
+  hasDecimalData?: boolean;
+  hasNegativeData?: boolean;
+}