+++ projects/fhi-angular-components/CHANGELOG.md
@@ -1,4 +1,11 @@
-# 4.2.0
+# 4.3.0
+
+> Jun 13, 2024
+
+* :tada: **Enhancement** `minDate` and `maxDate` inputs added to `FhiDateTimeComponent`
+* :tada: **Enhancement** Default `maxYear` and `maxDate` is set to `2099` and `31.12.2099`, repectively
+
+## 4.2.0
 
 > May 31, 2024