+++ projects/fhi-angular-highcharts/CHANGELOG.md
@@ -1,4 +1,17 @@
-# 3.3.0
+# 4.0.0
+
+> Jun 13, 2024
+
+* :bug: **Bugfix** Isolate diagram type group state to each instance of `FhiAngularHighchartsComponent`
+* :tada: **Enhancement** Add support for showing units in the footer
+* :tada: **Enhancement** Add support for a new diagram type: "Dobbel akse, linje og søyle" (`columnAndLine`)
+* :boom: **Breaking** Rename `FhiDiagramOptions.unit` to `FhiDiagramOptions.units` since it's an array
+* :boom: **Breaking** Diagram type ids for maps is changed to be in sync with maps from [code.highcharts.com](https://code.highcharts.com/mapdata/)
+  * `mapFylker` -> "Norway 2024" (was "Norway 2023" in v3.3.0)
+  * `mapFylker2023` -> "Norway 2023" (new)
+  * `mapFylker2019` -> "Norway 2019" (unchanged)
+
+## 3.3.0
 
 > May 31, 2024